--- a/src/main/kotlin/com/kneelawk/wiredredstone/part/RedrawablePart.kt
+++ b/src/main/kotlin/com/kneelawk/wiredredstone/part/RedrawablePart.kt
@@ -2,4 +2,6 @@ package com.kneelawk.wiredredstone.part
 
 interface RedrawablePart {
     fun redraw()
+
+    fun recalculateShape()
 }