--- a/src/main/kotlin/com/kneelawk/wiredredstone/util/ConnectableUtils.kt
+++ b/src/main/kotlin/com/kneelawk/wiredredstone/util/ConnectableUtils.kt
@@ -76,27 +76,63 @@ object ConnectableUtils {
         val side = part.side
         val pos = part.getSidedPos()
 
-        val nodes1 = net.getNodesAt(pos).filter { it.data.ext is ConnectablePartExt }
-            // The flatMap here causes entire parts to visually connect, even if only one of their network-nodes is
-            // actually connected.
-            .flatMap { node ->
-                node.connections.mapNotNull { link ->
+        val connections = net.getNodesAt(pos).filter { it.data.ext is ConnectablePartExt }
+            // The fold here causes entire parts to visually connect, even if only one of their network-nodes is
+            // actually connected. The first byte represents the actual connection value. The second byte keeps track of
+            // if this edge has already tried to connect in a different connection.
+            .fold(Pair(0u.toUByte(), 0u.toUByte())) { connections, node ->
+                var newConn = connections
+                node.connections.forEach { link ->
                     val other = link.other(node)
                     if (node.data.pos == other.data.pos && other.data.ext is SidedPartExt) {
-                        Connection(other.data.ext.side, ConnectionType.INTERNAL)
+                        newConn = setSingularConnection(
+                            newConn, blockage, side, other.data.ext.side, ConnectionUtils::setInternal,
+                            ConnectionUtils::isInternal
+                        )
                     } else {
-                        other.data.pos.subtract(node.data.pos.offset(side)).let { Direction.fromVector(it) }
-                            ?.let { Connection(it, ConnectionType.CORNER) } ?: other.data.pos.subtract(node.data.pos)
-                            .let { Direction.fromVector(it) }?.let { Connection(it, ConnectionType.EXTERNAL) }
+                        val cornerEdge = other.data.pos.subtract(node.data.pos.offset(side)).let(Direction::fromVector)
+                        if (cornerEdge != null) {
+                            newConn = setSingularConnection(
+                                newConn, blockage, side, cornerEdge, ConnectionUtils::setCorner,
+                                ConnectionUtils::isCorner
+                            )
+                        } else {
+                            other.data.pos.subtract(node.data.pos).let(Direction::fromVector)?.let {
+                                newConn = setSingularConnection(
+                                    newConn, blockage, side, it, ConnectionUtils::setExternal,
+                                    ConnectionUtils::isExternal
+                                )
+                            }
+                        }
                     }
                 }
-            }.groupBy { it.edge }.mapNotNull { (_, v) -> v.distinct().singleOrNull() }
+                newConn
+            }
 
-        val connections = nodes1.fold(0u.toUByte()) { current, new -> new.setForSide(side, current, blockage) }
-        val newConnections = part.overrideConnections(connections)
+        val newConnections = part.overrideConnections(connections.first)
 
         part.updateConnections(newConnections)
         (part as? RedrawablePart)?.redraw()
+    }
+
+    private inline fun setSingularConnection(
+        connections: Pair<UByte, UByte>, blockage: UByte, side: Direction, edge: Direction,
+        set: (UByte, Direction) -> UByte, test: (UByte, Direction) -> Boolean
+    ): Pair<UByte, UByte> {
+        val cardinal = RotationUtils.unrotatedDirection(side, edge)
+
+        if (!DirectionUtils.isHorizontal(cardinal) || BlockageUtils.isBlocked(blockage, cardinal)) {
+            return connections
+        }
+
+        return if (!ConnectionUtils.isDisconnected(connections.second, cardinal) && !test(
+                connections.second, cardinal
+            )
+        ) {
+            Pair(ConnectionUtils.setDisconnected(connections.first, cardinal), connections.second)
+        } else {
+            Pair(set(connections.first, cardinal), set(connections.second, cardinal))
+        }
     }
 
     /**