--- a/src/main/kotlin/com/kneelawk/wiredredstone/util/ConnectableUtils.kt
+++ b/src/main/kotlin/com/kneelawk/wiredredstone/util/ConnectableUtils.kt
@@ -17,21 +17,6 @@ import net.minecraft.util.shape.VoxelShape
 import net.minecraft.world.BlockView
 
 object ConnectableUtils {
-    data class Connection(val edge: Direction, val type: ConnectionType) {
-        fun setForSide(side: Direction, connections: UByte, blockage: UByte = BlockageUtils.UNBLOCKED): UByte {
-            val cardinal = RotationUtils.unrotatedDirection(side, edge)
-            if (!DirectionUtils.isHorizontal(cardinal) || BlockageUtils.isBlocked(blockage, cardinal)) {
-                // A blockage or a vertical connection means we shouldn't set anything
-                return connections
-            }
-
-            return when (type) {
-                ConnectionType.INTERNAL -> ConnectionUtils.setInternal(connections, cardinal)
-                ConnectionType.EXTERNAL -> ConnectionUtils.setExternal(connections, cardinal)
-                ConnectionType.CORNER -> ConnectionUtils.setCorner(connections, cardinal)
-            }
-        }
-    }
 
     /**
      * Updates both the connections and the blockage of the LMP part and asks it to update the client.
@@ -112,7 +97,10 @@ object ConnectableUtils {
         val newConnections = part.overrideConnections(connections.first)
 
         part.updateConnections(newConnections)
-        (part as? RedrawablePart)?.redraw()
+        (part as? RedrawablePart)?.let {
+            it.redraw()
+            it.recalculateShape()
+        }
     }
 
     private inline fun setSingularConnection(
@@ -132,6 +120,31 @@ object ConnectableUtils {
             Pair(ConnectionUtils.setDisconnected(connections.first, cardinal), connections.second)
         } else {
             Pair(set(connections.first, cardinal), set(connections.second, cardinal))
+        }
+    }
+
+    fun shouldUpdateConnectionsForNeighborUpdate(
+        shapeCache: MutableMap<Direction, VoxelShape>, world: ServerWorld, yourPos: BlockPos, otherPos: BlockPos
+    ): Boolean {
+        val direction = Direction.fromVector(otherPos.subtract(yourPos))
+        return if (direction != null) {
+            val previousShape = shapeCache[direction]
+            val state = world.getBlockState(otherPos)
+            val curShape = state.getOutlineShape(world, otherPos)
+
+            if (previousShape != null) {
+                // Identity checking here is about the best we can do if we want to be fast. This works because most
+                // blocks re-use the same VoxelShape objects if they want to keep their shapes.
+                if (previousShape !== curShape) {
+                    shapeCache[direction] = curShape
+                    true
+                } else false
+            } else {
+                shapeCache[direction] = curShape
+                true
+            }
+        } else {
+            true
         }
     }