# This PR
This pr provides several significant optimizations, mainly:
 * Diodes cause less block-updates.
 * Connection-recalculation is more efficient.
 * Network ref-rebuilds are more efficient.
 * Network connection attempts are more efficient.
 * Block-updates only cause network connection updates if the updating block's outline has changed.
 * Reduce the frequency of parts' outline-shape recalculations.

# Testing
I have tested this PR in two different redstone worlds and have found server tick times, that were previously unacceptably high, returning to acceptable levels. Previously I was seeing tick-times of 15 to 30ms, now tick times are 2 to 14ms.